% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leekasso.R
\name{leekasso}
\alias{leekasso}
\title{Building a Model with Top 10 Features}
\usage{
leekasso(x, y)
}
\arguments{
\item{x}{a n x p matrix of n observations and p predictors}

\item{y}{a vector of length n representing the response}
}
\value{
a vector of coefficients from the final fitted model with top 10 features
}
\description{
This function develops a prediction algorithm base on top 10 features
in 'x' that are most predictive of 'y'
}
\details{
This function run a univariate regression of 'y' on each predictor in 'x'
and calculates de p-value indicating the significance of the association.
The final set of 10 predictors is taken from the 10 smallest p-values.
}
\seealso{
\code{lm}
}
\author{
Juan Beleño
}
