% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leekasso.R
\name{predict10}
\alias{predict10}
\title{Prediction with the Top 10 Features}
\usage{
predict10(X, b)
}
\arguments{
\item{X}{a n x 10 matrix containing n new observation}

\item{b}{a vector of coefficients obtained from \code{leekasso} function}
}
\value{
a numeric vector containig the predicted values
}
\description{
This function takes a set of coefficients produced by the \code{leekasso}
function and makes the predictions for each of the values provided in the
input 'X' matrix.
}
